+++ packages/ci/src/lib/models.ts
@@ -19,6 +19,7 @@ export type Options = {
   debug?: boolean;
   detectNewIssues?: boolean;
   logger?: Logger;
+  skipComment?: boolean;
 };
 
 /**
@@ -27,11 +28,11 @@ export type Options = {
 export type Settings = Required<Options>;
 
 /**
- * Branches/commits for {@link runInCI}
+ * Branches/tags for {@link runInCI}
  */
 export type GitRefs = {
-  head: GitBranch;
-  base?: GitBranch;
+  head: string | GitBranch;
+  base?: string | GitBranch;
 };
 
 /**